refactor: Move snackbar logic into `LanguageScreen` and `MainScreen`
Move the `SnackbarHostState` and its related logic from `MainActivity` into the `LanguageScreen` and `MainScreen` composable.

This change encapsulates the snackbar functionality within the screen where it is used, making the component more self-contained. `LanguageScreen` now manages its own `SnackbarHostState` and displays a snackbar when a language change operation fails, removing the need to pass down `snackbarHostState` and `onShowSnackbar` callbacks from parent composables.

### Key Changes:
- **`LanguageScreen.kt`**:
  - Added a `SnackbarHost` and a locally managed `SnackbarHostState`.
  - The `onShowSnackbar` parameter has been removed, as the screen now handles showing its own snackbars.
- **`MainScreen.kt` & `MainActivity.kt`**:
  - Removed the `SnackbarHostState` and the `onShowSnackbar` lambda, simplifying the composable's signature.
- **Tests**:
  - `LanguageScreenSideEffectsTest` was updated to verify that the snackbar appears on the screen rather than checking if a callback was invoked.
  - `LanguageScreenContentTest` was updated to provide the newly required `snackbarHostState` parameter.
