+++ app/src/main/java/timur/gilfanov/messenger/MainActivity.kt
@@ -4,16 +4,12 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.navigation3.runtime.entry
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberNavBackStack
 import androidx.navigation3.ui.NavDisplay
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 import timur.gilfanov.messenger.domain.entity.chat.toChatId
 import timur.gilfanov.messenger.domain.entity.chat.toParticipantId
 import timur.gilfanov.messenger.navigation.Chat
@@ -47,20 +43,15 @@ class MainActivity : ComponentActivity() {
 @Suppress("LongMethod") // todo keep entities in feature modules
 @Composable
 fun MessengerApp() {
-    val scope = rememberCoroutineScope()
-
     val currentUserId = "550e8400-e29b-41d4-a716-446655440000".toParticipantId()
 
+    @Suppress("KotlinConstantConditions")
     val initBackStack = if (BuildConfig.FEATURE_SETTINGS) Main else ChatList
     val backStack = rememberNavBackStack(initBackStack)
 
     val onAuthFailure: () -> Unit = {
         backStack.clear()
         backStack.add(Login)
-    }
-    val snackbarHostState = remember { SnackbarHostState() }
-    val onShowSnackbar: (String) -> Unit = { message ->
-        scope.launch { snackbarHostState.showSnackbar(message) }
     }
 
     NavDisplay(
@@ -91,10 +82,8 @@ fun MessengerApp() {
             }
             entry<Main> {
                 MainScreen(
-                    snackbarHostState = snackbarHostState,
                     currentUserId = currentUserId,
                     onAuthFailure = onAuthFailure,
-                    onShowSnackbar = onShowSnackbar,
                     onChatClick = { chatId -> backStack.add(Chat(chatId.id.toString())) },
                     onNewChatClick = {
                         // Navigation to new chat screen will be implemented later
@@ -114,7 +103,6 @@ fun MessengerApp() {
             entry<Language> {
                 LanguageScreen(
                     onAuthFailure = onAuthFailure,
-                    onShowSnackbar = onShowSnackbar,
                     onBackClick = { backStack.removeLastOrNull() },
                 )
             }
