+++ app/src/main/java/timur/gilfanov/messenger/ui/screen/settings/LanguageScreen.kt
@@ -14,15 +14,20 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.MediumTopAppBar
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
+import kotlinx.coroutines.launch
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 import timur.gilfanov.messenger.R
@@ -33,13 +38,18 @@ import timur.gilfanov.messenger.ui.theme.MessengerTheme
 @Composable
 fun LanguageScreen(
     onAuthFailure: () -> Unit,
-    onShowSnackbar: (String) -> Unit,
     onBackClick: () -> Unit,
     modifier: Modifier = Modifier,
     viewModel: LanguageViewModel = hiltViewModel(),
 ) {
     val uiState by viewModel.collectAsState()
+    val snackbarHostState = remember { SnackbarHostState() }
+    val scope = rememberCoroutineScope()
     val errorMessage = stringResource(R.string.settings_language_change_failed)
+
+    val onShowSnackbar: (String) -> Unit = { message ->
+        scope.launch { snackbarHostState.showSnackbar(message) }
+    }
 
     viewModel.collectSideEffect {
         when (it) {
@@ -50,6 +60,7 @@ fun LanguageScreen(
 
     LanguageScreenContent(
         uiState = uiState,
+        snackbarHostState = snackbarHostState,
         onSelectLanguage = viewModel::changeLanguage,
         onBackClick = onBackClick,
         modifier = modifier,
@@ -60,12 +71,14 @@ fun LanguageScreen(
 @Composable
 fun LanguageScreenContent(
     uiState: LanguageUiState,
+    snackbarHostState: SnackbarHostState,
     onSelectLanguage: (UiLanguage) -> Unit,
     onBackClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Scaffold(
         modifier = modifier.fillMaxSize(),
+        snackbarHost = { SnackbarHost(snackbarHostState) },
         topBar = {
             MediumTopAppBar(
                 title = { Text(stringResource(R.string.settings_language_screen_title)) },
@@ -143,6 +156,7 @@ private fun Content(darkTheme: Boolean) {
     MessengerTheme(darkTheme = darkTheme) {
         LanguageScreenContent(
             uiState = uiState,
+            snackbarHostState = remember { SnackbarHostState() },
             onSelectLanguage = {},
             onBackClick = {},
         )
